# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RankPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRequest {
  /** The content to analyze */
  content: string;
}

/**
 * Category this suggestion belongs to
 */
export type SuggestionCategory =
  (typeof SuggestionCategory)[keyof typeof SuggestionCategory];

export const SuggestionCategory = {
  SEO: "SEO",
  AEO: "AEO",
  GEO: "GEO",
} as const;

export interface Suggestion {
  /** Short title of the suggestion */
  title: string;
  /** Detailed explanation of the suggestion */
  explanation: string;
  /** Category this suggestion belongs to */
  category: SuggestionCategory;
}

export interface AnalyzeResponse {
  /** SEO score from 0 to 100 */
  seoScore: number;
  /** AEO score from 0 to 100 */
  aeoScore: number;
  /** GEO score from 0 to 100 */
  geoScore: number;
  /** List of identified issues */
  issues: string[];
  /** List of suggestions with explanations */
  suggestions: Suggestion[];
}

export interface OptimizeRequest {
  /** The content to optimize */
  content: string;
}

export interface OptimizeResponse {
  /** The optimized version of the content */
  optimizedContent: string;
}

export interface ErrorResponse {
  error: string;
}