# lib/api-zod/src/generated/types/analyzeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RankPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Suggestion } from "./suggestion";

export interface AnalyzeResponse {
  /** SEO score from 0 to 100 */
  seoScore: number;
  /** AEO score from 0 to 100 */
  aeoScore: number;
  /** GEO score from 0 to 100 */
  geoScore: number;
  /** List of identified issues */
  issues: string[];
  /** List of suggestions with explanations */
  suggestions: Suggestion[];
}