# lib/api-zod/src/generated/types/optimizeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RankPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OptimizeRequest {
  /** The content to optimize */
  content: string;
}