# lib/api-zod/src/generated/types/optimizeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RankPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OptimizeResponse {
  /** The optimized version of the content */
  optimizedContent: string;
}