# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RankPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionCategory } from "./suggestionCategory";

export interface Suggestion {
  /** Short title of the suggestion */
  title: string;
  /** Detailed explanation of the suggestion */
  explanation: string;
  /** Category this suggestion belongs to */
  category: SuggestionCategory;
}